+++ src/server/exceptions/EncryptionException.js
@@ -1,0 +1,15 @@
+/*
+ * This file is part of the agenda-display project.
+ *
+ * (c) Nils Bohrs
+ *
+ * For the full copyright and license information, please view the LICENSE.md
+ * file that was distributed with this source code.
+ */
+'use strict';
+
+export class EncryptionException extends Error {
+    constructor(message) {
+        super(message);
+    }
+}