+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/bohrsty/agenda-display/compare/v0.1.0...HEAD)
 ### Added
+- password field to calendar setting
 - color picker
 - colors in environment, config and global state
 - settings ui (timeout and caldav connections)
@@ -18,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - api endpoint for agenda (data mocked)
 
 ### Changed
+- api endpoint to encrypt password in redis
+- optimized error handling using api calls in ui
 - caldav connection form to add color
 - config api endpoint to deliver calendar colors
 - jsdoc to reflect function type correctly in contexts